Updated Delete,Insert,View,Update

+++ Result-Processing System/Project_v2/Class 1/class1_delete.cs
@@ -19,40 +19,94 @@ namespace Project_v2
             InitializeComponent();
         }
 
-        private void btn_Delete_data_Click(object sender, EventArgs e)
+        private void button1_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("Do you want to Delete the row", "Sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+            if (txt_roll.Text != "")
+            {
+                DialogResult result = MessageBox.Show("Do you want to Delete the row", "Sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    conn.Open();
+                    OleDbCommand cmd = new OleDbCommand();
+                    cmd.Connection = conn;
+                    int roll = int.Parse(txt_roll.Text);
+                    cmd.CommandText = "Delete from Std_Data where Roll_no='" + roll + "'";
+                    cmd.ExecuteNonQuery();
+                    MessageBox.Show("Data Deleted", "Deleted", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    txt_roll.Clear();
+                    conn.Close(); 
+                }
+            }
+            else
+            {
+                MessageBox.Show("Empty Field", "Error", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+            }
+
+            
+        }
+
+        private void button2_Click(object sender, EventArgs e)
+        {
+             if (txt_num.Text!="")
+            {
+                DialogResult result = MessageBox.Show("Do you want to Delete the row", "Sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    conn.Open();
+                    OleDbCommand cmd = new OleDbCommand();
+
+                    int roll = int.Parse(txt_num.Text);
+                    cmd.Connection = conn;
+                    cmd.CommandText = "Delete from Class_1 where roll_no=" + roll + "";
+                    cmd.ExecuteNonQuery();
+                    MessageBox.Show("Data Deleted", "Deleted", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    txt_num.Clear();
+                    conn.Close();
+                }
+            }
+            else
+             {
+                 MessageBox.Show("Empty Field", "Error", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+             }
+            
+
+        }
+
+        private void button3_Click(object sender, EventArgs e)
+        {
+            DialogResult result = MessageBox.Show("Do you want to clean the Database?","ERROR", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
             if (result == DialogResult.Yes)
             {
                 conn.Open();
                 OleDbCommand cmd = new OleDbCommand();
                 cmd.Connection = conn;
-                int roll = int.Parse(txt_roll.Text);
-                cmd.CommandText = "Delete from Std_Data where Roll_no='" + roll + "'";
+                cmd.CommandText = "Delete from Std_Data";
                 cmd.ExecuteNonQuery();
-                MessageBox.Show("Data Deleted");
+                MessageBox.Show("Database Emptyed", "Emptyed", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 conn.Close();
             }
-
         }
 
-        private void btn_delete_Click(object sender, EventArgs e)
+        private void button4_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("Do you want to Delete the row", "Sure?", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+            DialogResult result = MessageBox.Show("Do you want to clean the Database?","ERROR", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
             if (result == DialogResult.Yes)
             {
                 conn.Open();
                 OleDbCommand cmd = new OleDbCommand();
                 cmd.Connection = conn;
-                int roll = int.Parse(txt_num.Text);
-                cmd.CommandText = "Delete from Std_Data where roll_no='" + roll + "'";
+                cmd.CommandText = "Delete from Class_1";
                 cmd.ExecuteNonQuery();
-                MessageBox.Show("Data Deleted","Deleted");
+                MessageBox.Show("Database Emptyed", "Emptyed", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 conn.Close();
             }
+        
         }
+
+
 
 
 
     }
 }
+
